Index creator for files
This module processes files and creates an index that informs
for each line of the file how many bytes are in the file where
the line ends.
For example, the for the last line we will have the byte size of
the file contents. For the first line we would have the byte size
of that line only.