Add line retrieve module
Give a line number this module will return the respective
line from the file. If the line is out of bounds it returns
an error.

First it consults the index repository to retrieve the necessary index
values to read the line.
Afterwards opens the file at the specific byte position the line starts
and reads the contents until the line ends.